cleaned up unsupported iteration
