+++ src/main/java/kodkod/engine/ExtendedSolver.java
@@ -5,7 +5,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Set;
 
 import kodkod.ast.Formula;
 import kodkod.ast.IntExpression;
